requirements.txt
absl-py==2.0.0
altgraph==0.17.2
asgiref==3.5.2
asttokens==2.4.1
attrs==23.1.0
bidict==0.22.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cvzone==1.6.1
cycler==0.12.1
Cython==3.0.5
decorator==5.1.1
distlib==0.3.4
Django==4.0.5
django-bootstrap3==21.2
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.7.1
filterpy==1.4.5
Flask==3.0.0
flask-ngrok==0.0.25
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
future==0.18.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
h11==0.14.0
idna==3.4
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
lapx==0.5.5
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mediapipe==0.10.7
mpmath==1.3.0
mysql-connector-python==8.2.0
networkx==3.2.1
numpy==1.26.0
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.2
parso==0.8.3
pefile==2021.9.3
Pillow==10.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.40
protobuf==4.21.12
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
PyJWT==2.4.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-engineio==4.8.0
python-socketio==5.3.0
pytz==2022.1
pywin32-ctypes==0.2.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
seaborn==0.13.0
sentry-sdk==1.34.0
simple-websocket==1.0.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.2
stack-data==0.6.3
supervision==0.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
tracker==0.1.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2022.1
ultralytics==8.0.208
urllib3==2.0.7
virtualenv==20.14.1
wcwidth==0.2.9
Werkzeug==3.0.1
wsproto==1.2.0
